requirements.txt
streamlit==0.89.0
tensorflow-gpu==2.2.0
opencv-python==4.5.3.56
numpy==1.19.5
scikit-image==0.16.2
scipy==1.4.1
h5py==2.10.0
keras==2.3.1
